requirements.txt
appdirs==1.4.3
Django==1.11
django-filter==1.0.2
djangorestframework==3.6.2
Markdown==2.6.8
numpy==1.12.1
olefile==0.44
packaging==16.8
pandas==0.20.1
Pillow==4.1.1
pkg-resources==0.0.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
six==1.10.0
SQLAlchemy==1.1.9
